@@ -1,5 +1,6 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
+import { HttpClientModule } from '@angular/common/http';
 import { HttpModule } from '@angular/http';
 
 import { FormsModule } from '@angular/forms';
@@ -28,12 +29,13 @@ import { AppRoutingModule } from './app.routing.module';
     CoursesModule,
     BrowserModule,
     FormsModule,
+    HttpModule,
     StoreModule.forRoot({
       courses: coursesReducer
     }),
     StoreDevtoolsModule.instrument(),
     AppRoutingModule,
-    HttpModule
+    HttpClientModule
   ],
   providers: [],
   bootstrap: [AppComponent]
